feat: Refactor frontend for improved code quality
This commit refactors the entire frontend to address several key issues:

- **No DRY Principle:** Centralized data fetching and component loading to eliminate redundant code.
- **No Centralized State:** Introduced a central store for menu data.
- **No Config or Environment Separation:** Moved hardcoded API endpoints to a configuration file.
- **No Documentation:** Added a README.md and AGENTS.md file.

The new structure is modular and more maintainable, using modern JavaScript features like ES modules.
