+++ get-in-touch.html
@@ -9,8 +9,7 @@
     <link href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined" rel="stylesheet"/>
     <script src="https://cdn.tailwindcss.com?plugins=forms,container-queries"></script>
     <link rel="stylesheet" href="style.css">
-    <script src="app.js" defer></script>
-    <script src="contact.js" defer></script>
+    <script type="module" src="src/app.js" defer></script>
 </head>
 <body class="contact-gradient-bg">
