+++ src/config.js
@@ -1,0 +1,5 @@
+export const config = {
+    api: {
+        menuUrl: 'menu.json'
+    }
+};
